Refactor: Input 파싱에 대한 책임을 도메인, 컨트롤러에서 모두 InputView로 이동
