+++ src/main/java/lotto/controller/LottoController.java
@@ -1,7 +1,6 @@
 package lotto.controller;
 
 import java.util.List;
-import java.util.stream.Collectors;
 import lotto.domain.Lotto;
 import lotto.domain.LottoResult;
 import lotto.domain.Lottos;
@@ -16,7 +15,7 @@ public class LottoController {
     PurchaseAmount purchaseAmount = InputView.getPurchaseAmount();
     int lottoCount = purchaseAmount.getLottoCount();
 
-    List<Lotto> manualLottos = getManualLottos(lottoCount);
+    List<Lotto> manualLottos = InputView.getManualLottos(lottoCount);
     Lottos lottos = Lottos.create(manualLottos, lottoCount);
     ResultView.printLottos(lottos);
 
@@ -24,20 +23,5 @@ public class LottoController {
 
     LottoResult result = lottos.calculate(lastWeekWinningLotto);
     ResultView.printResult(result, result.calculateYield(purchaseAmount));
-  }
-
-  private static List<Lotto> getManualLottos(int lottoCount) {
-    int manualLottoCount = InputView.getManualLottoCount();
-    validateCount(lottoCount, manualLottoCount);
-    List<String> manualInput = InputView.getManualLottoInput(manualLottoCount);
-    return manualInput.stream()
-        .map(Lotto::manualCreate)
-        .collect(Collectors.toList());
-  }
-
-  private static void validateCount(int lottoCount, int manualLottoCount) {
-    if (lottoCount < manualLottoCount) {
-      throw new IllegalArgumentException("수동 로또 갯수가 전체 로또 갯수보다 클 수 없습니다.");
-    }
   }
 }
